# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: LibraryOS API - Multi-tenant school bookstore management platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /shifts:
    get:
      operationId: listShifts
      tags: [shifts]
      summary: List shifts for current tenant
      parameters:
        - name: staffId
          in: query
          schema:
            type: integer
        - name: date
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: List of shifts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Shift"
  /shifts/open:
    post:
      operationId: openShift
      tags: [shifts]
      summary: Open a new shift
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenShiftRequest"
      responses:
        "201":
          description: Shift opened
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shift"
  /shifts/current:
    get:
      operationId: getCurrentShift
      tags: [shifts]
      summary: Get current open shift for logged-in cashier
      responses:
        "200":
          description: Current shift
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shift"
        "404":
          description: No open shift
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /shifts/{shiftId}/close:
    post:
      operationId: closeShift
      tags: [shifts]
      summary: Close a shift
      parameters:
        - name: shiftId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CloseShiftRequest"
      responses:
        "200":
          description: Shift closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shift"
  /shifts/{shiftId}:
    get:
      operationId: getShift
      tags: [shifts]
      summary: Get shift details with summary
      parameters:
        - name: shiftId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Shift details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShiftWithSummary"

  # ========== INVOICES ==========
...
  /staff:
    get:
      operationId: listStaff
      tags: [staff]
      summary: List staff for current tenant
      responses:
        "200":
          description: List of staff
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Staff"
    post:
      operationId: createStaff
      tags: [staff]
      summary: Create a staff member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStaffRequest"
      responses:
        "201":
          description: Staff created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Staff"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /staff/{staffId}:
    patch:
      operationId: updateStaff
      tags: [staff]
      summary: Update staff member
      parameters:
        - name: staffId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStaffRequest"
      responses:
        "200":
          description: Staff updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Staff"
    delete:
      operationId: deleteStaff
      tags: [staff]
      summary: Delete staff member
      parameters:
        - name: staffId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /staff/{staffId}/toggle-status:
    post:
      operationId: toggleStaffStatus
      tags: [staff]
      summary: Activate or deactivate staff account
      parameters:
        - name: staffId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Status toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Staff"

  # ========== DISCOUNTS ==========
...
  /reports/sales:
    get:
      operationId: getSalesReport
      tags: [reports]
      summary: Get sales report
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [daily, weekly, monthly]
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Sales report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesReport"
...
  /reports/top-products:
    get:
      operationId: getTopProducts
      tags: [reports]
      summary: Get top-selling products
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Top products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopProduct"
  /reports/payment-methods:
    get:
      operationId: getPaymentMethodsReport
      tags: [reports]
      summary: Get payment methods breakdown
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Payment methods report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentMethodsReport"

  # ========== NOTIFICATIONS ==========
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Shift:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        staffId:
          type: integer
        staffName:
          type: string
        openingBalance:
          type: number
        closingBalance:
          type: ["number", "null"]
        status:
          type: string
          enum: [open, closed]
        openedAt:
          type: string
          format: date-time
        closedAt:
          type: ["string", "null"]
          format: date-time
      required:
        - id
        - tenantId
        - staffId
        - staffName
        - openingBalance
        - status
        - openedAt
    ShiftWithSummary:
      allOf:
        - $ref: "#/components/schemas/Shift"
        - type: object
          properties:
            totalInvoices:
              type: integer
            totalSales:
              type: number
            cashSales:
              type: number
            cardSales:
              type: number
            totalReturns:
              type: number
          required:
            - totalInvoices
            - totalSales
            - cashSales
            - cardSales
            - totalReturns
    OpenShiftRequest:
      type: object
      properties:
        openingBalance:
          type: number
      required:
        - openingBalance
    CloseShiftRequest:
      type: object
      properties:
        closingBalance:
          type: number
      required:
        - closingBalance

    # INVOICE
...
    Staff:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        name:
          type: string
        username:
          type: string
        isActive:
          type: boolean
        lastLoginAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - tenantId
        - name
        - username
        - isActive
        - createdAt
    CreateStaffRequest:
      type: object
      properties:
        name:
          type: string
        username:
          type: string
        password:
          type: string
      required:
        - name
        - username
        - password
    UpdateStaffRequest:
      type: object
      properties:
        name:
          type: string
        username:
          type: string
        password:
          type: string

    # DISCOUNT
...
    SalesReport:
      type: object
      properties:
        period:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/SalesDataPoint"
        totalSales:
          type: number
        totalInvoices:
          type: integer
        categoryBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/CategorySalesBreakdown"
      required:
        - period
        - data
        - totalSales
        - totalInvoices
        - categoryBreakdown
    SalesDataPoint:
      type: object
      properties:
        date:
          type: string
        sales:
          type: number
        invoices:
          type: integer
      required:
        - date
        - sales
        - invoices
    CategorySalesBreakdown:
      type: object
      properties:
        category:
          type: string
        sales:
          type: number
        percentage:
          type: number
      required:
        - category
        - sales
        - percentage
...
    TopProduct:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        category:
          type: string
        quantitySold:
          type: integer
        revenue:
          type: number
      required:
        - productId
        - productName
        - category
        - quantitySold
        - revenue
    PaymentMethodsReport:
      type: object
      properties:
        cashTotal:
          type: number
        cardTotal:
          type: number
        cashPercent:
          type: number
        cardPercent:
          type: number
      required:
        - cashTotal
        - cardTotal
        - cashPercent
        - cardPercent

    # NOTIFICATION
...